pyproject.toml
[project]
name = "qa-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "camelot-py>=1.0.0",
    "fastapi>=0.116.1",
    "fastparquet>=2024.11.0",
    "granian>=2.5.0",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "pyarrow>=21.0.0",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "mypy",
    "ruff",
    "types-fpdf2",
]

[tool.ruff]
target-version = "py313"
fix = true
unsafe-fixes = true
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "EM",
    "FBT",
    "TRY003",
    "D1",
    "D203",
    "D213",
    "G004",
    "FA",
    "COM812",
    "ISC001",
]

[tool.ruff.lint.isort]
no-lines-before = ["standard-library", "local-folder"]
known-third-party = []
known-local-folder = []
lines-after-imports = 2

[tool.ruff.lint.extend-per-file-ignores]
"tests/*.py" = ["S101", "S311"]

[tool.pytest.ini_options]
addopts = '--cov=. -p no:warnings --cov-report term-missing --ignore=tests/test_migrations.py'

[tool.coverage.report]
exclude_also = ["if typing.TYPE_CHECKING:"]

[tool.coverage.run]
concurrency = ["thread", "greenlet"]
omit = ["*/migrations/*", "tests/test_migrations.py", "bin/*"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]
